Extract DeyeInverterState class
